+++ server/entry-node.ts
@@ -14,10 +14,10 @@ import { type DeployPlatform, setDeployPlatform } from './runtime-platform'
 import { syncPendingCloudTrafficReports } from './usecases/cloud-traffic-metering'
 import { buildCloudInstanceInfo, runtimeInfo } from './usecases/instance-info'
 import { INSTANCE_TELEMETRY_CRON, reportInstanceTelemetry } from './usecases/instance-telemetry'
-import { runLicensingRefresh } from './usecases/licensing-refresh-runner'
+import { runLicensingRefresh } from './usecases/licensing'
+import { purgeExpiredTrash, resolveTrashRetentionDays } from './usecases/purge'
 import { syncPendingRemoteDownloadUsageReports } from './usecases/remote-download-usage'
 import { getSitePublicOrigin } from './usecases/site-public-origin'
-import { purgeExpiredTrash, resolveTrashRetentionDays } from './usecases/trash-retention'
 
 const REFRESH_INTERVAL_MS = 6 * 60 * 60 * 1000 // 6 hours
 const TRAFFIC_SYNC_INTERVAL_MS = 10 * 60 * 1000 // 10 minutes
