+++ server/http/licensing.ts
@@ -7,9 +7,7 @@ import { originFromRequestUrl } from '../domain/site-public-origin'
 import type { Env } from '../middleware/platform'
 import { syncPendingCloudTrafficReports } from '../usecases/cloud-traffic-metering'
 import { buildCloudInstanceInfo, runtimeInfo } from '../usecases/instance-info'
-import { normalizeHost } from '../usecases/license-certificate'
-import { loadBindingState } from '../usecases/licensing'
-import { runLicensingRefresh } from '../usecases/licensing-refresh-runner'
+import { loadBindingState, normalizeHost, runLicensingRefresh } from '../usecases/licensing'
 import { syncPendingRemoteDownloadUsageReports } from '../usecases/remote-download-usage'
 import { getSitePublicOrigin } from '../usecases/site-public-origin'
