+++ server/middleware/require-feature.ts
@@ -3,8 +3,7 @@ import type { Context } from 'hono'
 import { createMiddleware } from 'hono/factory'
 import { ZPAN_CLOUD_URL_DEFAULT } from '../../shared/constants'
 import { hasFeature } from '../domain/licensing'
-import { normalizeHost } from '../usecases/license-certificate'
-import { loadBindingState } from '../usecases/licensing'
+import { loadBindingState, normalizeHost } from '../usecases/licensing'
 import { getSitePublicOrigin } from '../usecases/site-public-origin'
 import type { Env } from './platform'
