+++ workers/scheduled.ts
@@ -5,9 +5,9 @@ import { createDeps } from '../server/composition'
 import { createCloudflarePlatform } from '../server/platform/cloudflare'
 import { syncPendingCloudTrafficReports } from '../server/usecases/cloud-traffic-metering'
 import { INSTANCE_TELEMETRY_CRON, reportInstanceTelemetry } from '../server/usecases/instance-telemetry'
-import { runLicensingRefresh } from '../server/usecases/licensing-refresh-runner'
+import { runLicensingRefresh } from '../server/usecases/licensing'
+import { purgeExpiredTrash, resolveTrashRetentionDays } from '../server/usecases/purge'
 import { syncPendingRemoteDownloadUsageReports } from '../server/usecases/remote-download-usage'
-import { purgeExpiredTrash, resolveTrashRetentionDays } from '../server/usecases/trash-retention'
 import { ZPAN_CLOUD_URL_DEFAULT } from '../shared/constants'
 
 // Subset of the worker Env used by the scheduled handler.
